# docs/source/specs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "The API for Role Based Access Control.",
    "version": "1.0.0",
    "title": "Role Based Access Control",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://opensource.org/licenses/AGPL-3.0"
    }
  },
  "security": [
    {
      "basic_auth": []
    }
  ],
...
  "paths": {
...
    "/access/": {
      "get": {
        "tags": [
          "Access"
        ],
        "summary": "Get the permitted access for a principal in the tenant (defaults to principal from the identity header)",
        "description": "Access responses are sorted in ascending order by an ID internal to the database",
        "operationId": "getPrincipalAccess",
        "parameters": [
          {
            "name": "application",
            "in": "query",
            "description": "The application name(s) to obtain access for the principal. This is an exact match. When no application is supplied, all permissions for the principal are returned. You may also use a comma-separated list to match on multiple applications.",
            "required": true,
            "allowEmptyValue": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "username",
            "in": "query",
            "description": "Unique username of the principal to obtain access for (only available for admins, and if supplied, takes precedence over the identity header).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "description": "Parameter for ordering roles by value. For inverse ordering, supply '-' before the param value, such as: ?order_by=-application",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "application",
                "resource_type",
                "verb"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Set the status of users to get back.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "enabled",
                "disabled",
                "all"
              ],
              "default": "enabled"
            }
          },
          {
            "$ref": "#/components/parameters/QueryLimit"
          },
          {
            "$ref": "#/components/parameters/QueryOffset"
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of access objects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessPagination"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://console.redhat.com/{basePath}",
      "description": "Production Server",
      "variables": {
        "basePath": {
          "default": "api/rbac/v1"
        }
      }
    },
    {
      "url": "https://console.stage.redhat.com/{basePath}",
      "description": "Staging Server",
      "variables": {
        "basePath": {
          "default": "api/rbac/v1"
        }
      }
    },
    {
      "url": "http://localhost:{port}/{basePath}",
      "description": "Development Server",
      "variables": {
        "port": {
          "default": "8000"
        },
        "basePath": {
          "default": "api/rbac/v1"
        }
      }
    }
  ],
  "components": {
    "parameters": {
      "QueryOffset": {
        "in": "query",
        "name": "offset",
        "required": false,
        "description": "Parameter for selecting the offset of data.",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0
        }
      },
      "QueryLimit": {
        "in": "query",
        "name": "limit",
        "required": false,
        "description": "Parameter for selecting the amount of data returned.",
        "schema": {
          "type": "integer",
          "default": 10,
          "minimum": 1,
          "maximum": 1000
        }
      },
...
    },
...
    "securitySchemes": {
      "basic_auth": {
        "type": "http",
        "description": "The userid/password is needed when accessing this API externally",
        "scheme": "basic"
      }
    },
    "schemas": {
      "Error": {
        "required": [
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string",
                  "example": "Not found."
                },
                "status": {
                  "type": "string",
                  "example": "403"
                }
              }
            }
          }
        }
      },
...
      "PaginationMeta": {
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64",
            "example": 30
          }
        }
      },
      "PaginationLinks": {
        "properties": {
          "first": {
            "type": "string",
            "format": "uri",
            "example": "/api/v1/(resources)/?offset=0&limit=10"
          },
          "previous": {
            "type": "string",
            "format": "uri",
            "example": "/api/v1/(resources)/?offset=20&limit=10"
          },
          "next": {
            "type": "string",
            "format": "uri",
            "example": "/api/v1/(resources)/?offset=40&limit=10"
          },
          "last": {
            "type": "string",
            "format": "uri",
            "example": "/api/v1/(resources)/?offset=90&limit=10"
          }
        }
      },
      "ListPagination": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "links": {
            "$ref": "#/components/schemas/PaginationLinks"
          }
        }
      },
...
      "ResourceDefinitionFilter": {
          "oneOf": [
              {
                "$ref": "#/components/schemas/ResourceDefinitionFilterOperationEqual"
              },
              {
                "$ref": "#/components/schemas/ResourceDefinitionFilterOperationIn"
              }
          ]
     },
      "ResourceDefinitionFilterOperationEqual": {
        "type": "object",
        "required": [
            "key",
            "operation",
            "value"
        ],
        "properties": {
          "key": {
            "type": "string",
            "example": "cost-management.aws.account"
          },
          "operation": {
            "type": "string",
            "enum": [
              "equal"
            ]
          },
          "value": {
            "type": "string",
            "example": "123456",
            "nullable": true
          }
        }
      },
      "ResourceDefinitionFilterOperationIn": {
        "type": "object",
        "required": [
            "key",
            "operation",
            "value"
        ],
        "properties": {
          "key": {
            "type": "string",
            "example": "cost-management.aws.account"
          },
          "operation": {
            "type": "string",
            "enum": [
              "in"
            ]
          },
          "value": {
            "type": "array",
            "items": {
              "type": "string",
              "nullable": true
            },
            "example": ["value1", null, "value2"]
          }
        }
      },
      "ResourceDefinition": {
        "required": [
          "attributeFilter"
        ],
        "properties": {
          "attributeFilter": {
            "$ref": "#/components/schemas/ResourceDefinitionFilter"
          }
        }
      },
      "Access": {
        "required": [
          "permission",
          "resourceDefinitions"
        ],
        "properties": {
          "permission": {
            "type": "string",
            "example": "cost-management:*:read"
          },
          "resourceDefinitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceDefinition"
            }
          }
        }
      },
...
      "AccessPagination": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ListPagination"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Access"
                }
              }
            }
          }
        ]
      },
...
    }
  }
}